Refactor TextSplitterService to separate concerns and improve structure
